requirements.txt
aiohappyeyeballs          2.4.3
aiohttp                   3.9.5
aiosignal                 1.3.1
annotated-types           0.7.0
anyio                     4.6.0
attrs                     24.2.0
blinker                   1.8.2
certifi                   2024.8.30
charset-normalizer        3.3.2
click                     8.1.7
colorama                  0.4.6
dataclasses-json          0.6.7
fastapi                   0.115.0
filelock                  3.16.1
Flask                     3.0.3
Flask-Cors                5.0.0
frozenlist                1.4.1
fsspec                    2024.9.0
greenlet                  3.1.1
h11                       0.14.0
httpcore                  1.0.6
httpx                     0.27.2
huggingface-hub           0.25.1
idna                      3.10
itsdangerous              2.2.0
Jinja2                    3.1.4
joblib                    1.4.2
jsonpatch                 1.33
jsonpointer               3.0.0
langchain                 0.3.2
langchain-community       0.3.1
langchain-core            0.3.9
langchain-huggingface     0.1.0
langchain-pinecone        0.2.0
langchain-text-splitters  0.3.0
langsmith                 0.1.131
MarkupSafe                2.1.5
marshmallow               3.22.0
mpmath                    1.3.0
multidict                 6.1.0
mypy-extensions           1.0.0
networkx                  3.3
numpy                     1.26.4
orjson                    3.10.7
packaging                 24.1
pillow                    10.4.0
pinecone-client           5.0.1
pinecone-plugin-inference 1.1.0
pinecone-plugin-interface 0.0.7
pip                       24.2
pydantic                  2.9.2
pydantic_core             2.23.4
pydantic-settings         2.5.2
python-dotenv             1.0.1
PyYAML                    6.0.2
regex                     2024.9.11
requests                  2.32.3
requests-toolbelt         1.0.0
safetensors               0.4.5
scikit-learn              1.5.2
scipy                     1.14.1
sentence-transformers     3.1.1
setuptools                75.1.0
sniffio                   1.3.1
SQLAlchemy                2.0.35
starlette                 0.38.6
sympy                     1.13.3
tenacity                  8.5.0
threadpoolctl             3.5.0
tokenizers                0.20.0
torch                     2.4.1
tqdm                      4.66.5
transformers              4.45.1
typing_extensions         4.12.2
typing-inspect            0.9.0
urllib3                   2.2.3
uvicorn                   0.31.0
Werkzeug                  3.0.4
yarl                      1.13.1